requirements.txt
langchain
langchain-community
langchain-tavily
langgraph
fastapi
qdrant-client
python-dotenv